Moved state functions and locks into StateManager
